Implementando configurações diferentes do AWSSQS para profile prod e test
