+++ application/src/main/java/com/fiap/burger/application/config/AwsSQSConfigurationTest.java
@@ -1,0 +1,34 @@
+package com.fiap.burger.application.config;
+
+import com.amazonaws.services.sqs.AmazonSQSAsync;
+import com.amazonaws.services.sqs.AmazonSQSAsyncClientBuilder;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.cloud.aws.messaging.config.annotation.EnableSqs;
+import org.springframework.cloud.aws.messaging.core.QueueMessagingTemplate;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+import org.springframework.context.annotation.Profile;
+
+@Profile("test")
+@EnableSqs
+@Configuration
+public class AwsSQSConfigurationTest
+{
+    @Value("${cloud.aws.region}")
+    private String awsRegion;
+    @Value("${cloud.aws.end-point.uri}")
+    private String awsEndpointUri;
+
+    @Bean
+    @Primary
+    public AmazonSQSAsync amazonSQSAsync() {
+        return AmazonSQSAsyncClientBuilder.defaultClient();
+    }
+
+    @Bean
+    @Primary
+    public QueueMessagingTemplate queueMessagingTemplate(AmazonSQSAsync amazonSQSAsync) {
+        return new QueueMessagingTemplate(amazonSQSAsync);
+    }
+}
